fix: pull TestServer into a separate file
